peppi::frame
Module immutable

Immutable (fully-parsed) frame data, as Arrow arrays.

This is what you get when you parse a game in one shot using crate::io::slippi::read or crate::io::peppi::read.

These arrays can be shared, and cloning them is O(1). See the arrow2 docs for more.

Structs

Data

Frame data for a single character (ICs are two characters).

DreamlandWhispy

This event only occurs on Dreamland 64, and is sent whenever Whispy changes blow directions.

End

Information about the end of the game.

FodPlatform

This event only occurs on Fountain of Dreams, and is sent for each change in platform height. If both platforms are moving, there will be two events per frame.

Frame

All frame data for a single game, in struct-of-arrays format.

Item

An active item (includes projectiles).

ItemMisc

Miscellaneous item state.

PortData

Frame data for a single port.

Position

2D position.

Post

Post-frame update data, for making decisions about game states (such as computing stats).

Pre

Pre-frame update data, required to reconstruct a replay.

StadiumTransformation

This event only occurs on Pokemon Stadium, and is sent whenever the transformation event or transformation type changes.

Start

Initialization data such as game mode, settings, characters & stage.

StateFlags

Miscellaneous state flags.

TriggersPhysical

Trigger state.

Velocities

Self-induced and knockback velocities.

Velocity

2D velocity.