peppi::frame::immutable
Struct Frame

pub struct Frame {
    pub id: PrimitiveArray<i32>,
    pub ports: Vec<PortData>,
    pub start: Option<Start>,
    pub end: Option<End>,
    pub item: Option<Item>,
    pub item_offset: Option<OffsetsBuffer<i32>>,
    pub fod_platform: Option<FodPlatform>,
    pub fod_platform_offset: Option<OffsetsBuffer<i32>>,
    pub dreamland_whispy: Option<DreamlandWhispy>,
    pub dreamland_whispy_offset: Option<OffsetsBuffer<i32>>,
    pub stadium_transformation: Option<StadiumTransformation>,
    pub stadium_transformation_offset: Option<OffsetsBuffer<i32>>,
}

All frame data for a single game, in struct-of-arrays format.

Fields
id: PrimitiveArray<i32>

Frame IDs start at -123 and increment each frame. May repeat in case of rollbacks

ports: Vec<PortData>

Port-specific data

start: Option<Start>

Start-of-frame data

end: Option<End>

End-of-frame data

item: Option<Item>

Item data

item_offset: Option<OffsetsBuffer<i32>>

Logically, each frame has its own array of items. But we represent all item data in a flat array, with this field indicating the start of each sub-array

fod_platform: Option<FodPlatform>fod_platform_offset: Option<OffsetsBuffer<i32>>dreamland_whispy: Option<DreamlandWhispy>dreamland_whispy_offset: Option<OffsetsBuffer<i32>>stadium_transformation: Option<StadiumTransformation>stadium_transformation_offset: Option<OffsetsBuffer<i32>>
Implementations

impl Frame

pub fn into_struct_array( self, version: Version, ports: &[PortOccupancy], ) -> StructArray

pub fn from_struct_array(array: StructArray, version: Version) -> Self

impl Frame

pub fn write<W: Write>(&self, w: &mut W, version: Version) -> Result<()>

impl Frame

pub fn len(&self) -> usize

pub fn transpose_one(&self, i: usize, version: Version) -> Frame

pub fn rollbacks(&self, keep: Rollbacks) -> Vec<bool>

Frames IDs may appear multiple times due to rollbacks. This fn lets you “dedupe” rollbacks, by returning true for all but one of each unique frame ID. The value returned at index i corresponds to self.id[i].