peppi::frame::immutable
Struct Post

pub struct Post {

    pub character: PrimitiveArray<u8>,
    pub state: PrimitiveArray<u16>,
    pub position: Position,
    pub direction: PrimitiveArray<f32>,
    pub percent: PrimitiveArray<f32>,
    pub shield: PrimitiveArray<f32>,
    pub last_attack_landed: PrimitiveArray<u8>,
    pub combo_count: PrimitiveArray<u8>,
    pub last_hit_by: PrimitiveArray<u8>,
    pub stocks: PrimitiveArray<u8>,
    pub state_age: Option<PrimitiveArray<f32>>,
    pub state_flags: Option<StateFlags>,
    pub misc_as: Option<PrimitiveArray<f32>>,
    pub airborne: Option<PrimitiveArray<u8>>,
    pub ground: Option<PrimitiveArray<u16>>,
    pub jumps: Option<PrimitiveArray<u8>>,
    pub l_cancel: Option<PrimitiveArray<u8>>,
    pub hurtbox_state: Option<PrimitiveArray<u8>>,
    pub velocities: Option<Velocities>,
    pub hitlag: Option<PrimitiveArray<f32>>,
    pub animation_index: Option<PrimitiveArray<u32>>,
    pub last_hit_by_instance: Option<PrimitiveArray<u16>>,
    pub instance_id: Option<PrimitiveArray<u16>>,
    pub validity: Option<Bitmap>,

}

Post-frame update data, for making decisions about game states (such as computing stats).

Information is collected at the end of collision detection, which is the last consideration of the game engine.

Fields
character: PrimitiveArray<u8>

In-game character (can only change for Zelda/Sheik)

state: PrimitiveArray<u16>

Character’s action state

position: Position

Character’s position

direction: PrimitiveArray<f32>

Direction the character is facing

percent: PrimitiveArray<f32>

Damage taken (percent)

shield: PrimitiveArray<f32>

Size/health of shield

last_attack_landed: PrimitiveArray<u8>

Last attack ID that this character landed

combo_count: PrimitiveArray<u8>

Combo count (as defined by the game)

last_hit_by: PrimitiveArray<u8>

Port that last hit this player. Bugged in Melee: will be set to 6 in certain situations

stocks: PrimitiveArray<u8>

Number of stocks remaining

state_age: Option<PrimitiveArray<f32>>

Added: v0.2 Number of frames action state has been active. Can have a fractional component

state_flags: Option<StateFlags>

Added: v2.0 State flags

misc_as: Option<PrimitiveArray<f32>>

Added: v2.0 Used for different things. While in hitstun, contains hitstun frames remaining

airborne: Option<PrimitiveArray<u8>>

Added: v2.0 Is the character airborne?

ground: Option<PrimitiveArray<u16>>

Added: v2.0 Ground ID the character last touched

jumps: Option<PrimitiveArray<u8>>

Added: v2.0 Number of jumps remaining

l_cancel: Option<PrimitiveArray<u8>>

Added: v2.0 L-cancel status (0 = none, 1 = successful, 2 = unsuccessful)

hurtbox_state: Option<PrimitiveArray<u8>>

Added: v2.1 Hurtbox state (0 = vulnerable, 1 = invulnerable, 2 = intangible)

velocities: Option<Velocities>

Added: v3.5 Self-induced and knockback velocities

hitlag: Option<PrimitiveArray<f32>>

Added: v3.8 Hitlag frames remaining

animation_index: Option<PrimitiveArray<u32>>

Added: v3.11 Animation the character is in

last_hit_by_instance: Option<PrimitiveArray<u16>>

Added: v3.16 Instance ID of the player/item that last hit this player

instance_id: Option<PrimitiveArray<u16>>

Added: v3.16 Unique, serial ID for each new action state across all characters. Resets to 0 on death

validity: Option<Bitmap>

Indicates which indexes are valid (None means “all valid”). Invalid indexes can occur on frames where a character is absent (ICs or 2v2 games)

Implementations

impl Post

pub fn transpose_one(&self, i: usize, version: Version) -> Post