peppi
Module game

A single game of Super Smash Brothers Melee.

The mutable/immutable distinction is essentially an artifact of the underlying Arrow library. You’ll only encounter mutable data if you’re parsing live games.

Modules

immutable

Immutable (fully-parsed) game data.

mutable

Mutable (in-progress) game data.

shift_jis

Melee’s internal string encoding.

Structs

Bytes

Container for raw bytes of Start & End events.

End

Information about the end of the game.

GeckoCodes

Binary blob of Gecko codes in use.

Match

Information about the match a game belongs to.

Netplay

Netplay name, connect code, and Slippi UID.

Player

Information about each player such as character, team, stock count, etc.

PlayerEnd

Player placements.

Quirks

Slippi quirks that we need to track for round-trip integrity.

Scene

Major & minor scene numbers.

Start

Information used to initialize the game such as the game mode, settings, characters & stage.

Team

Information about the team a player belongs to.

Ucf

Information about the “Universal Controller Fix” mod.

Enums

DashBack

Dashback fix type.

EndMethod

How the game ended.

Language

The language the game is set to.

PlayerType

How a player is controlled.

Port

A slot that can be occupied by a player.

ShieldDrop

Shield drop fix type.

Constants

ICE_CLIMBERS

Since ICs are unique mechanically, sometimes we need to treat them specially.

MAX_PLAYERS

Some modes allow more characters than ports, e.g. Cruel Melee.

NUM_PORTS

How many ports the game supports.

Traits

Game

Functions

port_occupancy